+++ src/documents/item-sheet-weapon.mjs
@@ -1,5 +1,5 @@
 import { CASTLE_FALKENSTEIN } from "../config.mjs";
-import CastleFalkenstein from "../castle-falkenstein.mjs";
+import { CastleFalkenstein } from "../castle-falkenstein.mjs";
 
 /**
  * Sheet for the 'weapon' item type.