+++ src/forms/perform-feat.mjs
@@ -1,8 +1,8 @@
-import CastleFalkenstein from "../castle-falkenstein.mjs";
+import { CastleFalkenstein } from "../castle-falkenstein.mjs";
 import { CASTLE_FALKENSTEIN } from "../config.mjs";
 
 // A form for performing a Feat.
-export default class CastleFalkensteinPerformFeat extends FormApplication {
+export class CastleFalkensteinPerformFeat extends FormApplication {
 
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {